Changed float to double

+++ Galaxy_Code/random.h
@@ -1,9 +1,9 @@
 #ifndef RANDOM_H
 #define RANDOM_H
 
-float random_uniform(float min, float max);
-float random_gaussian(float mean, float sigma);
-float sample_plummer_radius(float a);
-Vector3 random_point_on_sphere(float r);
+double random_uniform(double min, double max);
+double random_gaussian(double mean, double sigma);
+double sample_plummer_radius(double a);
+Vector3 random_point_on_sphere(double r);
 
 #endif
